Implement book deletion

+++ app/src/main/java/com/udayrana/uday_bookapp/MainActivity.kt
@@ -8,6 +8,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.udayrana.uday_bookapp.databinding.ActivityMainBinding
 import kotlinx.coroutines.launch
 
@@ -22,25 +23,16 @@ class MainActivity : AppCompatActivity(), BookClickListener {
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        bookList =
-            mutableListOf()
+        bookList = mutableListOf()
         bookAdapter = BookAdapter(bookList, this)
         binding.recyclerViewBookList.adapter = bookAdapter
         binding.recyclerViewBookList.layoutManager = LinearLayoutManager(this)
         binding.recyclerViewBookList.addItemDecoration(
-            DividerItemDecoration(
-                this,
-                LinearLayoutManager.VERTICAL
-            )
+            DividerItemDecoration(this, LinearLayoutManager.VERTICAL)
         )
-
         bookDao = AppDatabase.getInstance(this).bookDao()
-
+        binding.buttonAddBook.setOnClickListener { addBook() }
         refreshBookRecyclerView()
-
-        binding.buttonAddBook.setOnClickListener {
-            addBook()
-        }
     }
 
     override fun onResume() {
@@ -59,11 +51,7 @@ class MainActivity : AppCompatActivity(), BookClickListener {
     }
 
     private fun addBook() {
-        startActivity(Intent(this@MainActivity, AddBookActivity::class.java))
-    }
-
-    override fun displayBookDetails(book: Book) {
-        TODO("Not yet implemented")
+        startActivity(Intent(this@MainActivity, BookFormActivity::class.java))
     }
 
     override fun editBook(book: Book) {
@@ -71,12 +59,28 @@ class MainActivity : AppCompatActivity(), BookClickListener {
     }
 
     override fun deleteBook(book: Book) {
-        lifecycleScope.launch {
-            bookDao.delete(book)
-            Toast.makeText(this@MainActivity, "Deleted $book.title", Toast.LENGTH_LONG)
-                .show()
-            refreshBookRecyclerView()
+        MaterialAlertDialogBuilder(this@MainActivity).setTitle("Confirm deletion")
+            .setMessage("Are you sure you want to delete ${book.title} by ${book.author}?")
+            .setNegativeButton("Cancel") { dialog, _ ->
+                dialog.cancel()
+            }
+            .setPositiveButton("Delete") { dialog, _ ->
+                lifecycleScope.launch {
+                    bookDao.delete(book)
+                    Toast.makeText(
+                        this@MainActivity,
+                        "Deleted ${book.title} by ${book.author}",
+                        Toast.LENGTH_LONG
+                    )
+                        .show()
+                    refreshBookRecyclerView()
+                }
+                dialog.dismiss()
+            }
+            .show()
+    }
 
-        }
+    override fun displayBookDetails(book: Book) {
+        TODO("Not yet implemented")
     }
 }
